More changes than I can recall
The window initialization has been simplified. The window creation
interface takes a configuration object instead of individual parameters
for each setting, because of how difficult it can be to spot signature
mismatches at the moment.

A lot fewer abbreviations are used now.

Most of the constants in the 'wrapper' module have been split into several
different enumerations.

The hint defaults are now the same as those for GLFW.

The interface supports GLFW 3.2 now.